Update directory for gamefi
