+++ src/views/Gamefi/components/InputComponent.tsx
